refactor: Separate view files
