+++ ASBInterviewExercise/Transaction/Presentation/Views/TransactionListView.swift
@@ -31,59 +31,3 @@ struct TransactionListView: View {
         }
     }
 }
-
-struct TransactionRow: View {
-    let transaction: Transaction
-    
-    var body: some View {
-        HStack {
-            Text(transaction.summary)
-            Spacer()
-            Text(transaction.debit > 0 ? "-\(transaction.debit, specifier: "%.2f")" : "+\(transaction.credit, specifier: "%.2f")")
-        }
-    }
-}
-
-struct TransactionDetailView: View {
-    @ObservedObject var viewModel: TransactionDetailViewViewModel
-
-    var body: some View {
-        VStack(alignment: .leading) {
-            Text("Summary: \(viewModel.transaction.summary)")
-                .font(.title)
-            Text("Date: \(viewModel.transaction.transactionDate))")
-            Text("Debit: \(viewModel.transaction.debit, specifier: "%.2f")")
-            Text("Credit: \(viewModel.transaction.credit, specifier: "%.2f")")
-            Spacer()
-        }
-        .padding()
-        .navigationTitle("Transaction Detail")
-    }
-}
-
-let transactionEntities: [TransactionEntity] = [
-     TransactionEntity(id: 1, transactionDate: "2021-08-31T15:47:10", summary: "Hackett, Stamm and Kuhn", debit: 9379.55, credit: 0),
-     TransactionEntity(id: 2, transactionDate: "2022-02-17T10:44:35", summary: "Hettinger, Wilkinson and Kshlerin", debit: 3461.35, credit: 0),
-     TransactionEntity(id: 3, transactionDate: "2021-02-21T08:19:12", summary: "McKenzie, Bins and Macejkovic", debit: 0, credit: 1415.74),
-     TransactionEntity(id: 4, transactionDate: "2021-12-15T14:35:20", summary: "Treutel, Schaden and Pfannerstill", debit: 1639.43, credit: 0)
- ]
-
-let transactionsFromEntities: [Transaction] = transactionEntities.map { Transaction(from: $0) }
-
-// Preview for TransactionRow
-struct TransactionRow_Previews: PreviewProvider {
-    static var previews: some View {
-        Group {
-            TransactionRow(transaction: transactionsFromEntities[0])
-            TransactionRow(transaction: transactionsFromEntities[2])
-        }
-        .previewLayout(.sizeThatFits) // Important for list row previews
-    }
-}
-
-// Preview for TransactionList
-//struct TransactionList_Preview: PreviewProvider {
-//    static var previews: some View {
-//        TransactionListView(transactionList: transactionsFromEntities)
-//    }
-//}
